requirements.txt
tensorflow
keras
numpy
Pillow
